Optimizations based on the properties of rare numbers
If A = 4 then Q = 0 and B - P = positive or negative Even digit i.e. -8, -6, -4, -2, 0, 2, 4, 6, 8.
If A = 6 then Q = 0 or 5 and B - P = positive or negative Odd digit i.e. -9, -7, -5, -3, -1, 1, 3, 5, 7, 9.
If A = 8 then Q = 2, 3, 7 or 8; if Q = 2 then B + P = 9, if Q = 3 then B - P = 7 for B > P and B - P = -3 for B < P and B can never be equal to P, if Q = 7 then B + P = 11 for B >1 and B + P =1 for B < 1, if Q = 8 then B = P.